third_party/blink/web_tests/external/wpt/navigation-api/ordering-and-transition/anchor-download-intercept_no-currententrychange-expected.txt

This is a testharness.js-based test.
[FAIL] event and promise ordering for <a download> intercepted by intercept()
  assert_array_equals: expected property 2 to be "navigatesuccess" but got "promise microtask" (expected array ["navigate", "handler run", "navigatesuccess", "promise microtask", "transition.finished fulfilled"] got ["navigate", "handler run", "promise microtask", "navigatesuccess", "transition.finished fulfilled"])
Harness: the test ran to completion.
